jacobian introduced

+++ data/dataset.py
@@ -44,7 +44,7 @@ class BOXDataset(Dataset):
     width = W: W ^|^| OX
     """
     def __init__(self, data_file, transform=None, train_val_mask=None):
-        self.data = torch.load(data_file)
+        self.data = torch.load(data_file).float()
         self.transform = transform
         self.bottom_boundary = bottom_slice(self.data)
         self.z_max, self.y_max, self.x_max = self.data.shape[-3:]
